Modification cloud-servicechain-state.yang key
Due to elan-service-chain-state is an augmentation of
elan-instance, elan-instance-name can be removed from
elan-to-pseudo-port-data list key. The new key will be
composed with elanLportTag and scfTag.

+++ vpnservice/cloud-servicechain/cloud-servicechain-impl/src/main/java/org/opendaylight/netvirt/cloudservicechain/ElanServiceChainHandler.java
@@ -56,7 +56,7 @@ public class ElanServiceChainHandler {
      *        ElanPseudoPort is not used in any other ServiceChain
      * @param addOrRemove States if the flows must be created or removed
      */
-    public void programElanScfPipeline(String elanName, short tableId, int scfTag, int elanLportTag,
+    public void programElanScfPipeline(String elanName, short tableId, long scfTag, int elanLportTag,
                                        boolean isLastServiceChain, int addOrRemove) {
         logger.info("programElanScfPipeline:  elanName={}   scfTag={}   elanLportTag={}   lastSc={}   addOrRemove={}",
                     elanName, scfTag, elanLportTag, isLastServiceChain, addOrRemove);
